Yay, functional cli test!
